+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - added axis-aligned bounding box
 - added bounding volume hierarchy traversal
+- added morton code computation 
 
 ## [0.0.2] - 2021-07-15
 ### Changed